Move classes into game_logic file
